[blockchain] formatting for NFT contract

+++ apps/blockchain/contracts/NFT.sol
@@ -11,18 +11,22 @@ contract NFT is ERC721URIStorage {
     Counters.Counter private idCounter;
     address private immutable contractAddress;
 
-    event Minted (
+    event Minted(
         string name,
         string symbol,
-        uint indexed tokenId,
+        uint256 indexed tokenId,
         string tokenURI
     );
 
-    constructor(string memory name, string memory symbol, address _contractAddress) ERC721(name, symbol) {
+    constructor(
+        string memory name,
+        string memory symbol,
+        address _contractAddress
+    ) ERC721(name, symbol) {
         contractAddress = _contractAddress;
     }
 
-    function getCurrentId() public view returns (uint) {
+    function getCurrentId() public view returns (uint256) {
         return idCounter.current();
     }
